# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "有来管理后台 API",
        "description": "基于 ThinkPHP 8 的 RBAC 权限管理系统后端 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "开发环境"
        }
    ],
    "paths": {
...
        "/api/v1/roles": {
            "get": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "角色列表",
                "description": "分页查询角色列表",
                "operationId": "b43135a47dbed904b5e9fb29f38ae459",
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "post": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "新增角色",
                "description": "创建角色",
                "operationId": "0ca20b2e6aab72bd05c041a70af70407",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/v1/roles/{id}/form": {
            "get": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "获取角色表单数据",
                "description": "获取角色表单数据",
                "operationId": "9b787a68df96bb49f2b3091d7b430c93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "get": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "获取角色详情",
                "description": "获取角色详情",
                "operationId": "50bbe89ef69e37118d2efe976e3bea3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "put": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "修改角色",
                "description": "更新角色",
                "operationId": "33ac8e8b65d8fc6b67e845dbe5255eae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/roles/{ids}": {
            "delete": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "删除角色",
                "description": "批量删除角色",
                "operationId": "303731aeaca8f2feca249df9019e42f2",
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "角色ID，多个以英文逗号分割",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/roles/{id}/status": {
            "patch": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "修改角色状态",
                "description": "修改角色状态",
                "operationId": "8690d916e101efce47d4290c412b0298",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/roles/{id}/menus": {
            "get": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "获取角色菜单ID集合",
                "description": "获取角色的菜单ID集合",
                "operationId": "59b1963901d79deeddbce56744be0ad1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "put": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "分配角色菜单权限",
                "description": "角色分配菜单权限",
                "operationId": "6f6a2a0aa06e165f4024327679a61753",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/roles/{id}/depts": {
            "get": {
                "tags": [
                    "03.角色接口"
                ],
                "summary": "获取角色部门ID集合",
                "description": "获取角色的部门ID集合(自定义数据权限)",
                "operationId": "fca1b2ec85418bda011e838adb4fa26f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    },
...
}